move around files and directories
