+++ plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/extensions/world.kt
@@ -1,6 +1,5 @@
 package dev.betrix.superSmashMobsBrawl.extensions
 
-import dev.betrix.superSmashMobsBrawl.maps.SpawnPoint as MapSpawnPoint
 import dev.betrix.superSmashMobsBrawl.models.SpawnPoint
 import org.bukkit.Location
 import org.bukkit.World
@@ -14,9 +13,4 @@ fun World.location(spawnPoint: SpawnPoint): Location {
         spawnPoint.yaw ?: 0f,
         spawnPoint.pitch ?: 90f,
     )
-}
-
-fun World.location(spawnPoint: MapSpawnPoint): Location {
-    val pos = spawnPoint.position
-    return Location(this, pos.x, pos.y, pos.z, spawnPoint.yaw ?: 0f, spawnPoint.pitch ?: 90f)
 }
