--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/maps/blueForest.kt
@@ -1,15 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.maps
-
-val blueForestHub = defineMap {
-    name = "Blue Forest Hub"
-    id = "blue_forest"
-    description = "The main hub world for players to gather and prepare for matches"
-    voidLevel = 0
-    type = MapType.HUB
-    maxPlayers = null
-    worldBorderSize = 500.0
-
-    addCreator("PLACEHOLDER_UUID")
-
-    spawnPoints { at(-29.5, 58.0, 1.5, 90f, 0f) }
-}
