+++ plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/passives/BarragePassive.kt
@@ -1,6 +1,5 @@
-package dev.betrix.superSmashMobsBrawl.brawl.passives
+package dev.betrix.superSmashMobsBrawl.passives
 
-import dev.betrix.superSmashMobsBrawl.brawl.BrawlPassive
 import dev.betrix.superSmashMobsBrawl.events.Damager
 import dev.betrix.superSmashMobsBrawl.events.SmashDamageEvent
 import dev.betrix.superSmashMobsBrawl.events.SmashDamageType
@@ -31,9 +30,9 @@ class BarragePassive(player: Player) : BrawlPassive("barrage", player) {
             event<PlayerInteractEvent> {
                 if (
                     player != this@BarragePassive.player ||
-                    player.inventory.itemInMainHand.type != Material.BOW ||
-                    !player.inventory.contains(Material.ARROW) ||
-                    (action != Action.RIGHT_CLICK_AIR && action != Action.RIGHT_CLICK_BLOCK)
+                        player.inventory.itemInMainHand.type != Material.BOW ||
+                        !player.inventory.contains(Material.ARROW) ||
+                        (action != Action.RIGHT_CLICK_AIR && action != Action.RIGHT_CLICK_BLOCK)
                 ) {
                     return@event
                 }
@@ -101,16 +100,17 @@ class BarragePassive(player: Player) : BrawlPassive("barrage", player) {
                     )
 
                 val arrow =
-                    ArrowProjectile(player, "Barrage Arrow", 3.0, spread).onHitLivingEntity { entity,
-                                                                                              arrow ->
+                    ArrowProjectile(player, "Barrage Arrow", 3.0, spread).onHitLivingEntity {
+                        entity,
+                        arrow ->
                         arrow.projectile?.remove()
 
                         SmashDamageEvent(
-                            entity,
-                            Damager.LivingEntity(player),
-                            6.0,
-                            damageType = SmashDamageType.Projectile,
-                        )
+                                entity,
+                                Damager.LivingEntity(player),
+                                6.0,
+                                damageType = SmashDamageType.Projectile,
+                            )
                             .callEvent()
 
                         true
